+++ DarkCodex/MartialArt.cs
@@ -1,5 +1,6 @@
 ﻿using CodexLib;
 using Kingmaker.Blueprints.Classes;
+using Kingmaker.Blueprints.Classes.Prerequisites;
 using Kingmaker.Blueprints.Items.Weapons;
 using Kingmaker.Designers.Mechanics.Facts;
 using Kingmaker.ElementsSystem;
@@ -412,6 +413,13 @@ namespace DarkCodex
                 );
 
             Helper.AddCombatFeat(feat);
+        }
+
+        [PatchInfo(Severity.Fix, "Fix Wolverine Rage", "unlock Limitless Rage for Wolverine (Shifter)", false)]
+        public static void FixShifterWolverineRage()
+        {
+            Helper.Get<BlueprintFeature>("5cb58e6e406525342842a073fb70d068") //LimitlessRage
+                .GetComponent<PrerequisiteFeaturesFromList>().Add("6c4079c4c751405abc48a6dea8a78cee"); //ShifterWildShapeWolverineFeature
         }
     }
 }